+++ src/ros2/usb_cam_node.cpp
@@ -72,15 +72,15 @@ UsbCamNode::UsbCamNode(const rclcpp::NodeOptions & node_options)
   this->declare_parameter("pixel_format", "yuyv");
   this->declare_parameter("av_device_format", "YUV422P");
   this->declare_parameter("video_device", "/dev/video0");
-  this->declare_parameter("brightness", 50);  // 0-255, -1 "leave alone"
+  this->declare_parameter("brightness", -100);  // -64,64, -100 "leave alone"
   this->declare_parameter("contrast", -1);    // 0-255, -1 "leave alone"
   this->declare_parameter("saturation", -1);  // 0-255, -1 "leave alone"
   this->declare_parameter("sharpness", -1);   // 0-255, -1 "leave alone"
   this->declare_parameter("gain", -1);        // 0-100?, -1 "leave alone"
   this->declare_parameter("auto_white_balance", true);
   this->declare_parameter("white_balance", 4000);
-  this->declare_parameter("autoexposure", true);
-  this->declare_parameter("exposure", 100);
+  this->declare_parameter("auto_exposure", 1);
+  this->declare_parameter("exposure_time_absolute", 100);
   this->declare_parameter("autofocus", false);
   this->declare_parameter("focus", -1);  // 0-255, -1 "leave alone"
 
@@ -223,8 +223,8 @@ void UsbCamNode::get_params()
     {
       "camera_name", "camera_info_url", "frame_id", "framerate", "image_height", "image_width",
       "io_method", "pixel_format", "av_device_format", "video_device", "brightness", "contrast",
-      "saturation", "sharpness", "gain", "auto_white_balance", "white_balance", "autoexposure",
-      "exposure", "autofocus", "focus"
+      "saturation", "sharpness", "gain", "auto_white_balance", "white_balance", "auto_exposure",
+      "exposure_time_absolute", "autofocus", "focus"
     }
   );
 
@@ -270,10 +270,10 @@ void UsbCamNode::assign_params(const std::vector<rclcpp::Parameter> & parameters
       m_parameters.auto_white_balance = parameter.as_bool();
     } else if (parameter.get_name() == "white_balance") {
       m_parameters.white_balance = parameter.as_int();
-    } else if (parameter.get_name() == "autoexposure") {
-      m_parameters.autoexposure = parameter.as_bool();
-    } else if (parameter.get_name() == "exposure") {
-      m_parameters.exposure = parameter.as_int();
+    } else if (parameter.get_name() == "auto_exposure") {
+      m_parameters.auto_exposure = parameter.as_int();
+    } else if (parameter.get_name() == "exposure_time_absolute") {
+      m_parameters.exposure_time_absolute = parameter.as_int();
     } else if (parameter.get_name() == "autofocus") {
       m_parameters.autofocus = parameter.as_bool();
     } else if (parameter.get_name() == "focus") {
@@ -289,7 +289,7 @@ void UsbCamNode::assign_params(const std::vector<rclcpp::Parameter> & parameters
 void UsbCamNode::set_v4l2_params()
 {
   // set camera parameters
-  if (m_parameters.brightness >= 0) {
+  if (m_parameters.brightness >= -64 && m_parameters.brightness <= 64) {
     RCLCPP_INFO(this->get_logger(), "Setting 'brightness' to %d", m_parameters.brightness);
     m_camera->set_v4l_parameter("brightness", m_parameters.brightness);
   }
@@ -325,17 +325,10 @@ void UsbCamNode::set_v4l2_params()
   }
 
   // check auto exposure
-  if (!m_parameters.autoexposure) {
-    RCLCPP_INFO(this->get_logger(), "Setting 'exposure_auto' to %d", 1);
-    RCLCPP_INFO(this->get_logger(), "Setting 'exposure' to %d", m_parameters.exposure);
-    // turn down exposure control (from max of 3)
-    m_camera->set_v4l_parameter("exposure_auto", 1);
-    // change the exposure level
-    m_camera->set_v4l_parameter("exposure_absolute", m_parameters.exposure);
-  } else {
-    RCLCPP_INFO(this->get_logger(), "Setting 'exposure_auto' to %d", 3);
-    m_camera->set_v4l_parameter("exposure_auto", 3);
-  }
+  RCLCPP_INFO(this->get_logger(), "Setting 'auto_exposure' to %d", m_parameters.auto_exposure);
+  RCLCPP_INFO(this->get_logger(), "Setting 'exposure_time_absolute' to %d", m_parameters.exposure_time_absolute);
+  m_camera->set_v4l_parameter("auto_exposure", m_parameters.auto_exposure);
+  m_camera->set_v4l_parameter("exposure_time_absolute", m_parameters.exposure_time_absolute);
 
   // check auto focus
   if (m_parameters.autofocus) {